#110 Move makeReducer to reducers
