+++ packages/actions/src/index.js
@@ -3,7 +3,6 @@ export { prefixedValueMirror as makeActionTypes } from '@redux-tools/utils';
 export { default as makeActionCreator } from './makeActionCreator';
 export { default as makeConstantActionCreator } from './makeConstantActionCreator';
 export { default as makeSimpleActionCreator } from './makeSimpleActionCreator';
-export { default as makeReducer } from './makeReducer';
 
 export { default as isErrorAction } from './isErrorAction';
 export { default as isNotErrorAction } from './isNotErrorAction';
