+++ packages/reducers/src/index.js
@@ -1,4 +1,4 @@
 export { default as makeEnhancer, storeInterface } from './makeEnhancer';
-export { default as getStateByAction } from './getStateByAction';
-export { default as getStateByNamespace } from './getStateByNamespace';
 export { default as composeReducers } from './composeReducers';
+export { default as makeReducer } from './makeReducer';
+export { getStateByAction, getStateByNamespace } from '@redux-tools/namespaces';
